Add hostname from log4j appender
Previously adding the hostname could be done manually, but it was a
little tricky, especially for Spark applications.

If the hostname is successfully retrieved it is set in the 'properties'
of the log4j message, which is then created as a dynamic property in the
Solr index. In the future the 'hostname' should be added as a named
field to the Solr 'schema.xml'