Add a helper class `Metrics`
The Metrics class makes it easier to write metrics to Solr and documents
how it might be done. This implementation uses log4j, so it is not as performant as it could be.
It is required that an entire log message is written to Solr, instead of just the
metric by itself.