Switch to Monix Observables from Akka Streams
The original intention was to fix an error where the Akka stream would
not terminate when a stream failed (like when it overflowed). I was able
eventually to get the stream to fail and notify the application so the
application could exit. I was unable to get this to work with Akka,
hence the library change.

The Monix code looks essentially the same, but has fewer 'knobs', is
simpler,  and is
better documented. It is also a smaller library that has active support.

The stream failure and exit is handled by a callback. When the
default 'fail' mode is chosen for an overloaded stream, a method is
called that includes a call to `System.exit`. This behavior is tested in
a unit test.

There are also a few style changes included in the changeset.