Fix a bug where `id` is always empty string
By setting the id to empty string, Solr is not allowed to create unique
ids for inserts and so every insert becomes an update. The result is
that only 1 record is ever recorded.

The test that would have caught this was commented out at the time the
changes were committed. It was subsequently updated to fit the outcome
of the errored result.