Adds CDH6 functionality
- Builds against both Solr 4 and Solr 7 by using the CDH_VERSION environmental varaible. If CDH_VERSION 5 Solr 4 will
be chosen, otherwise Solr7. Solr 4 and Solr 7 are submodules. Even if the build isn't running against one of the versions
it will still be tested
- Made a SolrService that's implemented by Solr4 and Solr7 modules
- Made a SolrProvider object that can build SolrService implementations in the two modules
- Upgraded the sbt version for some errors I was having. In the end this might not have been necessary, but it shouldn't
hurt anything so am planning on leaving it
- Update deploy scripts to work with CDH 5 and 6
- Test both cdh 5 and cdh 6 builds in Travis
- Modify CollectionRoller.scala to change all aliases pointing to a collection before deleting it, which is required in Solr 7
- Disabled all integration tests. A lot of the tests are integration tests, and none of them have dependencies that aren't spun
up in the test itself. Maybe they should be separated out sometime, but there were only a few test in `it` folders anyways.