Initialize httpclient outside of constructor
Initializing it inside the constructor leads to a null pointer exception
when other httpclients on the classpath are picked up and they try to
log with slf4j. Other http clients should not be used because they are shaded,
but something about Spark on CDH6 is causing this to happen.